backend/services/hotel/requirements.txt
# Function dependencies, for example:
# package>=version
firebase-admin>=5.2.0
google-cloud-pubsub>=2.13.4
